Revisi Obat dan Pembelian Obat Suppliers

+++ app/Http/Controllers/ObatController.php
@@ -9,6 +9,8 @@ use Illuminate\Support\Facades\Auth;
 use App\Models\KategoriObat;
 use App\Http\Requests\ObatRequest;
 use App\Http\Requests\ObatUpdateRequest;
+use App\Models\Transaksi;
+use App\Models\ResepObatDetail;
 
 class ObatController extends Controller
 {
@@ -125,8 +127,14 @@ class ObatController extends Controller
      */
     public function destroy($id)
     {
+        $resepObat = ResepObatDetail::where('id_obat', $id)->get();
         $obat = Obat::findOrFail($id);
-        $obat->delete();
-        return redirect()->back()->with("success", "Hapus data berhasil");
+        if ($resepObat) {
+            return redirect()->back()->with("error", "Hapus data tidak berhasil, karena Obat sudah digunakan di Resep Obat dan Transaksi");
+        }
+        else {
+            $obat->delete();
+            return redirect()->back()->with("success", "Hapus data berhasil");
+        }
     }
 }
